Email funzionano

+++ utils/InvioPDFLaureando2.php
@@ -19,39 +19,51 @@ class InvioPDFLaureando2 {
     /**
      * @AssociationType ProspettoPDFLaureando2
      */
-    private CorsoDiLaurea $_cdl;
-    private string $_dataLaurea;
-
-    /**
-     * @access public
-     * @param int[] aMatricole
-     * @ParamType aMatricole int[]
-     */
+    private ?CorsoDiLaurea $_cdl = null;
+    private ?string $_dataLaurea = null;
 
     public function __construct()
     {
+        $this->_matricole = array();
         $json_content = file_get_contents(AccessoProspetti::pathAusiliario());
+        if (!$json_content)
+            return;
         $obj = json_decode($json_content, true);
+        if (!isset($obj))
+        {
+            return;
+        }
         $this->_matricole = $obj['matricole'];
         $this->_cdl = Configurazione::CorsiDiLaurea()[$obj['cdl']];
         $this->_dataLaurea = $obj['data_laurea'];
     }
     public function invioProspetti(int $max = PHP_INT_MAX) : array
     {
+        if (!isset($this->_cdl) || !isset($this->_dataLaurea)) {
+            return array();
+        }
         $inviati = array();
         for ($j = 0; $j < min(sizeof($this->_matricole), $max); $j++)
         {   
             try {
+                if ($this->_matricole[$j] === 0) {
+                    // Sto inviando alla commissione
+                    if ($this->inviaProspetto($this->_cdl->EmailCommissione))
+                    {
+                        $inviati[] = 'Commissione';
+                        $this->_matricole[$j] = -1;
+                    }
+                    continue;
+                }
                 $prospetto = new ProspettoPdfLaureando2($this->_matricole[$j], $this->_cdl, $this->_dataLaurea);
-                if ($this->inviaProspetto($prospetto->_carrieraLaureando))
-                {
+                if ($this->inviaProspetto($prospetto->_carrieraLaureando)) {
                     unlink(AccessoProspetti::pathLaureandoServer($this->_matricole[$j]));
-                    $inviati[] = $this->_matricole[$j];
-                    $this->_matricole[$j] = 0;
+                    $inviati[] = (int)$this->_matricole[$j];
+                    $this->_matricole[$j] = -1;
                 }
             } catch (Exception $ex) {};
         }
-        $this->_matricole = array_filter($this->_matricole, function (int $a) { return $a !== 0; });
+        $this->_matricole = array_values(array_filter($this->_matricole, function (int $a) { return $a >= 0; }));
         $this->AggiornaFile();
         return $inviati;
     }
@@ -61,7 +73,7 @@ class InvioPDFLaureando2 {
      * @ReturnType bool
      */
     public function inviaProspetto(
-        CarrieraLaureando2|CarrieraLaureandoInformatica2 $studente_carriera) : bool {
+        CarrieraLaureando2|CarrieraLaureandoInformatica2|string $destinatario) : bool {
 
         $messaggio = new PHPMailer();
         $messaggio->IsSMTP();
@@ -70,12 +82,21 @@ class InvioPDFLaureando2 {
         $messaggio->SMTPAuth = false;
         $messaggio->Port = 25;
 
-        $messaggio->From = 'no-reply-laureandosi@ing.unipi.it';
-        $messaggio->AddAddress($studente_carriera->_email);
+        $messaggio->From = 'noreply-laureandosi@ing.unipi.it';
+        $messaggio->CharSet = 'UTF-8';
+        if (is_string($destinatario)) {
+            $messaggio->AddAddress($destinatario);
+        } else {
+            $messaggio->AddAddress($destinatario->_email);
+        }
         $messaggio->Subject = 'Appello di laurea in Ing. TEST- indicatori per voto di laurea';
         $messaggio->Body = $this->_cdl->FormulaEmail;
 
-        $messaggio->AddAttachment(AccessoProspetti::pathLaureandoServer($studente_carriera->_matricola));
+        if (is_string($destinatario)) {
+            $messaggio->AddAttachment(AccessoProspetti::pathCommissioneServer());
+        } else {
+            $messaggio->AddAttachment(AccessoProspetti::pathLaureandoServer($destinatario->_matricola));
+        }
         
         $res = $messaggio->Send();
         $messaggio->smtpClose();
@@ -107,9 +128,13 @@ class InvioPDFLaureando2 {
         }
         return $res > 0;
     }
-    public static function Generate(array $matricole, string $cdl, string $data_laurea) : InvioPDFLaureando2|null
+    public static function Generate(array $matricole, string $cdl, string $data_laurea) : ?InvioPDFLaureando2
     {
-        if (!self::SaveFile($matricole, $cdl, $data_laurea))
+        $copy = (new ArrayObject(array_map("intval", $matricole)))->getArrayCopy();
+        if (count($copy) === 0 || end($copy) !== 0) {
+            $copy[] = 0; // 0 significa inviare alla commissione
+        }
+        if (!self::SaveFile($copy, $cdl, $data_laurea))
             return null;
         return new InvioPDFLaureando2();
     }
